feature Json to .txt newletter
